#5 Add UserExpenseRepository.
