+++ src/Api/Interfaces/IUserCategoryRepository.cs
@@ -1,0 +1,11 @@
+﻿using PersonalFinanceApp.Models;
+
+namespace PersonalFinanceApp.Interfaces
+{
+    public interface IUserCategoryRepository
+    {
+        void ResetUserCategories(int id);
+        void DeleteUserCategory(int userId, int categoryId);
+        Category CreateUserCategory(int userId, CreateCategoryDto createCategoryDto);
+    }
+}
