+++ src/Api/Interfaces/IUserCategoryService.cs
@@ -1,0 +1,11 @@
+﻿using PersonalFinanceApp.Models;
+
+namespace PersonalFinanceApp.Interfaces
+{
+    public interface IUserCategoryService
+    {
+        void ResetUserCategories(int id);
+        void DeleteUserCategory(int userId, int categoryId);
+        Category CreateUserCategory(int userId, CreateCategoryDto createCategoryDto);
+    }
+}
