+++ src/Api/Repository/UserDbRepository.cs
@@ -82,94 +82,6 @@ namespace PersonalFinanceApp.Repository
             _context.SaveChanges();
             
         }
-
-        public void ResetUserCategories(int id)
-        {
-            var user = _context.Users.Include(u => u.Categories).FirstOrDefault(u => u.Id == id);
-            
-            user.Categories = DefaultCategory.GetDefaultCategories().Select(c => new Category { Name = c.Name }).ToList();
-            _context.SaveChanges();
-        }
-
-        public void DeleteUserCategory(int userId, int categoryId)
-        {
-            var category = _context.Categories.FirstOrDefault(c => c.Id == categoryId);
-
-            _context.Categories.Remove(category);
-            _context.SaveChanges();
-        }
-
-        public Category CreateUserCategory(int userId, CreateCategoryDto createCategoryDto)
-        {
-            var user = _context.Users.Include(u => u.Categories).FirstOrDefault(u => u.Id == userId);
-
-            var category = new Category
-            {
-                Name = createCategoryDto.Name
-            };
-
-            user.Categories.Add(category);
-            _context.SaveChanges();
-
-            return category;
-        }
-
-        public Expense CreateUserExpense(int userId, int categoryId, CreateExpenseDto createExpenseDto)
-        {
-            var user = _context.Users.Include(u => u.Expenses).FirstOrDefault(u => u.Id == userId);
-
-            var expense = new Expense 
-            {
-                Amount = createExpenseDto.Amount,
-                CategoryId = categoryId,
-                Date = createExpenseDto.Date,
-                Description = createExpenseDto.Description
-            };
-
-            user.Expenses.Add(expense);
-            _context.SaveChanges();
-
-            return expense;
-        }
-
-        public void DeleteUserExpense(int userId, int expenseId)
-        {
-            var expense = _context.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            _context.Expenses.Remove(expense);
-            _context.SaveChanges();
-        }
-
-        public List<Expense> GetUserExpenses(int userId)
-        {
-            var user = _context.Users
-                .Include(e => e.Expenses)
-                .FirstOrDefault(u => u.Id == userId);
-
-            return user.Expenses;
-        }
-
-        public Expense ChangeExpenseCategory(int userId, int expenseId, int newCategoryId)
-        {
-            var expense = _context.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            expense.CategoryId = newCategoryId;
-
-            _context.SaveChanges();
-            return expense;
-        }
-
-        public Expense UpdateUserExpense(int userId, int expenseId, CreateExpenseDto createExpenseDto)
-        {
-            var expense = _context.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            expense.Amount = createExpenseDto.Amount;
-            expense.Date = createExpenseDto.Date;
-            expense.Description = createExpenseDto.Description;
-
-            _context.SaveChanges();
-            return expense;
-        }
     }
 }
