+++ src/Api/Services/UserService.cs
@@ -102,152 +102,5 @@ namespace PersonalFinanceApp.Services
 
             _userRepository.DeleteUser(id);
         }
-
-        public void ResetUserCategories(int id)
-        {
-            var user = _userRepository.GetUser(id);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            _userRepository.ResetUserCategories(id);
-        }
-
-        public void DeleteUserCategory(int userId, int categoryId)
-        {
-            var user = _userRepository.GetUser(userId);
-            var category = user.Categories.FirstOrDefault(c => c.Id == categoryId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            if (category == null)
-            {
-                throw new KeyNotFoundException("Category not found");
-            }
-
-            _userRepository.DeleteUserCategory(userId, categoryId);
-        }
-
-        public Category CreateUserCategory(int userId, CreateCategoryDto createCategoryDto)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            return _userRepository.CreateUserCategory(userId, createCategoryDto);
-        }
-
-        public Expense CreateUserExpense(int userId, int categoryId, CreateExpenseDto createExpenseDto)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            var category = user.Categories.FirstOrDefault(c => c.Id == categoryId);
-            
-            if (category == null)
-            {
-                throw new KeyNotFoundException("Category not found");
-            }
-
-            if (createExpenseDto.Amount <= 0)
-            {
-                throw new ArgumentException("Incorrect expense amount.");
-            }
-
-            return _userRepository.CreateUserExpense(userId, categoryId, createExpenseDto);
-        }
-
-        public void DeleteUserExpense(int userId, int expenseId)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            var expense = user.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            if (expense == null)
-            {
-                throw new KeyNotFoundException("Expense not found");
-            }
-
-            _userRepository.DeleteUserExpense(userId, expenseId);
-        }
-
-        public List<Expense> GetUserExpenses(int userId)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            return _userRepository.GetUserExpenses(userId);
-        }
-
-        public Expense ChangeExpenseCategory(int userId, int expenseId, int newCategoryId)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            var expense = user.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            if (expense == null)
-            {
-                throw new KeyNotFoundException("Expense not found");
-            }
-
-            var category = user.Categories.FirstOrDefault(c => c.Id == newCategoryId);
-
-            if (category == null)
-            {
-                throw new KeyNotFoundException("Category not found");
-            }
-
-            return _userRepository.ChangeExpenseCategory(userId, expenseId, newCategoryId);
-        }
-
-        public Expense UpdateUserExpense(int userId, int expenseId, CreateExpenseDto createExpenseDto)
-        {
-            var user = _userRepository.GetUser(userId);
-
-            if (user == null)
-            {
-                throw new KeyNotFoundException("User not found");
-            }
-
-            var expense = user.Expenses.FirstOrDefault(e => e.Id == expenseId);
-
-            if (expense == null)
-            {
-                throw new KeyNotFoundException("Expense not found");
-            }
-
-            if (createExpenseDto.Amount <= 0)
-            {
-                throw new ArgumentException("Incorrect expense amount.");
-            }
-
-            return _userRepository.UpdateUserExpense(userId, expenseId, createExpenseDto);
-        }
     }   
 }
